pyproject.toml
[tool.poetry]
name = "chordite"
version = "0.1.0"
description = "Chorded layout analyzer"
authors = ["Eve <eve@propter.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
